+++ backend/src/main/java/com/redhat/cloud/notifications/db/repositories/DrawerNotificationRepository.java
@@ -26,7 +26,7 @@ public class DrawerNotificationRepository {
     public Integer updateReadStatus(String orgId, String username, Set<UUID> notificationIds, Boolean readStatus) {
 
         String hql = "UPDATE DrawerNotification SET read = :readStatus "
-            + "WHERE id.orgId = :orgId and id.userId = :userId and id.event.id in (:notificationIds)";
+            + "WHERE id.orgId = :orgId and id.userId = :userId and id.eventId in (:notificationIds)";
 
         return entityManager.createQuery(hql)
             .setParameter("orgId", orgId)
@@ -41,9 +41,9 @@ public class DrawerNotificationRepository {
         query.setSortFields(DrawerNotification.SORT_FIELDS);
         query.setDefaultSortBy("created:DESC");
         Optional<Query.Sort> sort = query.getSort();
-        String hql = "SELECT dn.id.event.id, dn.read, " +
-            "dn.id.event.bundleDisplayName, dn.id.event.applicationDisplayName, dn.id.event.eventTypeDisplayName, dn.created, dn.id.event.renderedDrawerNotification, bundle.name "
-            + "FROM DrawerNotification dn join Bundle bundle on dn.id.event.bundleId = bundle.id where dn.id.orgId = :orgId and dn.id.userId = :userid";
+        String hql = "SELECT dn.id.eventId, dn.read, " +
+            "dn.event.bundleDisplayName, dn.event.applicationDisplayName, dn.event.eventTypeDisplayName, dn.created, dn.event.renderedDrawerNotification, bundle.name "
+            + "FROM DrawerNotification dn join Bundle bundle on dn.event.bundleId = bundle.id where dn.id.orgId = :orgId and dn.id.userId = :userid";
 
         hql = addHqlConditions(hql, bundleIds, appIds, eventTypeIds, startDate, endDate, readStatus);
         if (sort.isPresent()) {
@@ -100,14 +100,14 @@ public class DrawerNotificationRepository {
 
             // add org id as criteria on event table to allow usage of
             // index ix_event_org_id_bundle_id_application_id_event_type_display_nam
-            hql += " AND dn.id.event.orgId = :orgId";
+            hql += " AND dn.event.orgId = :orgId";
 
             if (eventTypesNotEmpty) {
-                hql += " AND dn.id.event.eventType.id IN (:eventTypeIds)";
+                hql += " AND dn.event.eventType.id IN (:eventTypeIds)";
             } else if (applicationsNotEmpty) {
-                hql += " AND dn.id.event.applicationId IN (:appIds)";
+                hql += " AND dn.event.applicationId IN (:appIds)";
             } else if (bundlesNotEmpty) {
-                hql += " AND dn.id.event.bundleId IN (:bundleIds)";
+                hql += " AND dn.event.bundleId IN (:bundleIds)";
             }
         }
